CRAS: Make shared structures architecture-independent.
This allows x86 clients to connect to x86-64 server, which is
important for crouton.

Get rid of size_t, long, and timespec, whose size depend on the
architecture. We also pack all structures to make sure alignment
is the same across platforms.

We assume that sizeof(enum) = sizeof(int) = 4, which is true on
x86/x86-64/arm with GCC (using default compilation flags).

Volume is now an uint32_t (was size_t), capture gain is now an
int32_t (was long): no difference on x86/arm, and 32-bit is more
than enough range on x86-64. Frame rate, number of channels, and
shm offsets are also converted from size_t to uint32_t.

We replace uses of timespec by a fixed-size cras_timespec, where
both fields are 32-bit signed integers (as on x86/arm, x86-64
uses 64-bit ints). All we care about in CRAS is differences
between 2 cras_timespec, so 64-bit fields are not required (even
after year 2038).

BUG=chromium:309961
TEST=Audio input/output still work.

+++ cras/src/common/cras_audio_format.c
@@ -32,8 +32,8 @@ static const int channel_alt[CRAS_CH_MAX][CRAS_CH_MAX] =
 
 /* Create an audio format structure. */
 struct cras_audio_format *cras_audio_format_create(snd_pcm_format_t format,
-						   size_t frame_rate,
-						   size_t num_channels)
+						   uint32_t frame_rate,
+						   uint32_t num_channels)
 {
 	struct cras_audio_format *fmt;
 	unsigned i;
@@ -79,7 +79,7 @@ void cras_audio_format_destroy(struct cras_audio_format *fmt)
 	free(fmt);
 }
 
-float** cras_channel_conv_matrix_alloc(size_t in_ch, size_t out_ch)
+float** cras_channel_conv_matrix_alloc(uint32_t in_ch, uint32_t out_ch)
 {
 	size_t i;
 	float **p;
@@ -99,7 +99,7 @@ alloc_err:
 	return NULL;
 }
 
-void cras_channel_conv_matrix_destroy(float **p, size_t out_ch)
+void cras_channel_conv_matrix_destroy(float **p, uint32_t out_ch)
 {
 	size_t i;
 	for (i = 0; i < out_ch; i ++)
